BREAKING CHANGE: Change default pull policy to IfNotPresent
Done as part of https://gitlab.com/gitlab-org/cluster-integration/auto-deploy-image/-/merge_requests/48 where we want to remove the current hardcoded IfNotPresent setting from auto-deploy-image and allow it to use these defaults. Therefore we need to align this file with the defaults that are currently in place. This value can be easily overridden see GitLab docs on Helm chart overrides for details.